🔀 :: (#68) 알림 읽음 여부 확인 API

+++ src/main/java/com/gsp/bridge/domain/notification/domain/NotificationListId.java
@@ -11,8 +11,8 @@ import java.io.Serializable;
 @EqualsAndHashCode
 public class NotificationListId implements Serializable {
 
-    private Long notificationEntity;
+    private Integer notificationEntity;
 
-    private Long company;
+    private Integer company;
 
 }
