+++ src/main/java/com/gsp/bridge/domain/notification/domain/repository/NotificationRepository.java
@@ -3,5 +3,5 @@ package com.gsp.bridge.domain.notification.domain.repository;
 import com.gsp.bridge.domain.notification.domain.NotificationEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NotificationRepository extends CrudRepository<NotificationEntity, Long> {
+public interface NotificationRepository extends CrudRepository<NotificationEntity, Integer> {
 }
